- added retrieval of next cpu instruction - changed data types of program counter and memory hash map key to int

+++ src/main/java/emulator/AddressRegister.java
@@ -8,13 +8,13 @@ package emulator;
 
 final class AddressRegister {
 
-    private short memory;
+    private int memory;
 
-    short read() {
+    int read() {
         return memory;
     }
 
-    void write(short value) {
+    void write(int value) {
         memory = value;
     }
